requirements.txt
flask
torch
torchvision
numpy
pillow
grad-cam
opencv-python-headless